Refactor SearchForm to be a ModelForm of Redditor
SearchForm is a form whose only entry is the username, which is the 
primary key of the Redditor model. Hence, it makes more sense for 
SearchForm to be derived from the model. In this way, when a custom 
validator for the username is implemented, it will be automatically 
applied to the form as well.