Fix search form validation bug
The SearchForm has been restored to a subclass of the Form class. This 
avoids the problem with the validation that was happening with the 
ModelForm class.

The validation of the username field in the Redditor model has been 
separated from the definition of the username field, so that it can be 
accessed for other uses (like the validation of the SearchForm).

A new test has been added: if the Redditor is already in the database, 
the SearchForm should be valid.