Refactor Redditor for client-side validation
The validation has been refactored in the Redditor model and form so 
that all parameters are specified once and not repeated.

The search form has been refactored so that now all extra classes are 
specified in the SearchForm class. This is needed to pass validation 
parameters to html for client-side validation with JavaScript.