Refactor JS validation to read input attributes from server
The client-side validator in search.js now reads regex and message from 
attributes of input.needs-validation tags. In principle, this is not 
restricted to just the search form but works with any input with the 
needs-validation class, but in our case the only application (at the 
moment) is to the search form.